requirements.txt
altair
streamlit
numpy
tensorflow
pillow